- New tests implemented in seperate assemblies to reduce impact on existing unittests
