+++ src/ILCompiler.MetadataTransform/src/ILCompiler/Metadata/IMetadataPolicy.cs
@@ -47,8 +47,6 @@ namespace ILCompiler.Metadata
         /// <summary>
         /// Return the Module that should be treated as defining the type. Typically implementations
         /// will return typeDef.Module, but in some circumstances it may return a different value.
-        /// The ModuleDesc that is returned will be cast to IAssemblyDesc and the GetName method is
-        /// the only function that shall be used on this ModuleDesc.
         /// </summary>
         Cts.ModuleDesc GetModuleOfType(Cts.MetadataType typeDef);
     }
