+++ src/ILCompiler.MetadataTransform/tests/SimpleTests.cs
@@ -29,14 +29,13 @@ namespace MetadataTransformTests
         public void TestAllTypes()
         {
             var policy = new SingleFileMetadataPolicy();
-            policy.Init();
             var transformResult = MetadataTransform.Run(policy, new[] { _systemModule });
 
-            Assert.Equal(2, transformResult.Scopes.Count());
+            Assert.Equal(1, transformResult.Scopes.Count());
             
             Assert.Equal(
                 _systemModule.GetAllTypes().Count(x => !policy.IsBlocked(x)),
-                transformResult.Scopes.First().GetAllTypes().Count() + transformResult.Scopes.Last().GetAllTypes().Count());
+                transformResult.Scopes.First().GetAllTypes().Count());
         }
 
         [Fact]
@@ -46,7 +45,6 @@ namespace MetadataTransformTests
             // up in the __ComObject interface list.
 
             var policy = new SingleFileMetadataPolicy();
-            policy.Init();
             var transformResult = MetadataTransform.Run(policy, new[] { _systemModule });
 
             Cts.MetadataType icastable = _systemModule.GetType("System.Private.CompilerServices", "ICastable");
@@ -67,7 +65,6 @@ namespace MetadataTransformTests
         public void TestStandaloneSignatureGeneration()
         {
             var policy = new SingleFileMetadataPolicy();
-            policy.Init();
 
             var transformResult = MetadataTransform.Run(policy, new[] { _systemModule });
 
@@ -92,21 +89,18 @@ namespace MetadataTransformTests
         public void TestSampleMetadataGeneration()
         {
             var policy = new SingleFileMetadataPolicy();
-            policy.Init();
 
             var sampleMetadataModule = _context.GetModuleForSimpleName("SampleMetadataAssembly");
             var transformResult = MetadataTransform.Run(policy,
                 new[] { _systemModule, sampleMetadataModule });
 
-            Assert.Equal(4, transformResult.Scopes.Count);
+            Assert.Equal(2, transformResult.Scopes.Count);
 
             var systemScope = transformResult.Scopes.Single(s => s.Name.Value == "PrimaryMetadataAssembly");
             var sampleScope = transformResult.Scopes.Single(s => s.Name.Value == "SampleMetadataAssembly");
-            var windowsWinRTScope = transformResult.Scopes.Single(s => s.Name.Value == "Windows");
-            var sampleWinRTScope = transformResult.Scopes.Single(s => s.Name.Value == "SampleMetadataWinRT");
 
-            Assert.Equal(_systemModule.GetAllTypes().Count(t => !policy.IsBlocked(t)), systemScope.GetAllTypes().Count() + windowsWinRTScope.GetAllTypes().Count());
-            Assert.Equal(sampleMetadataModule.GetAllTypes().Count(t => !policy.IsBlocked(t)), sampleScope.GetAllTypes().Count() + sampleWinRTScope.GetAllTypes().Count());
+            Assert.Equal(_systemModule.GetAllTypes().Count(t => !policy.IsBlocked(t)), systemScope.GetAllTypes().Count());
+            Assert.Equal(sampleMetadataModule.GetAllTypes().Count(t => !policy.IsBlocked(t)), sampleScope.GetAllTypes().Count());
 
             // TODO: check individual types
         }
@@ -119,11 +113,10 @@ namespace MetadataTransformTests
             var transformResult = MetadataTransform.Run(policy,
                 new[] { _systemModule, sampleMetadataModule });
 
-            Assert.Equal(2, transformResult.Scopes.Count);
+            Assert.Equal(1, transformResult.Scopes.Count);
 
             var sampleScope = transformResult.Scopes.First();
-            var sampleScopeWinRTThing = transformResult.Scopes.Last();
-            Assert.Equal(sampleMetadataModule.GetAllTypes().Count(t => !policy.IsBlocked(t)), sampleScope.GetAllTypes().Count() + sampleScopeWinRTThing.GetAllTypes().Count());
+            Assert.Equal(sampleMetadataModule.GetAllTypes().Count(t => !policy.IsBlocked(t)), sampleScope.GetAllTypes().Count());
 
             var objectType = (Cts.MetadataType)_context.GetWellKnownType(Cts.WellKnownType.Object);
             var objectRecord = transformResult.GetTransformedTypeReference(objectType);
@@ -192,7 +185,6 @@ namespace MetadataTransformTests
             Cts.MetadataType attributeHolder = sampleMetadataModule.GetType("BlockedMetadata", "AttributeHolder");
 
             var policy = new SingleFileMetadataPolicy();
-            policy.Init();
 
             var transformResult = MetadataTransform.Run(policy,
                 new[] { _systemModule, sampleMetadataModule });
@@ -219,107 +211,6 @@ namespace MetadataTransformTests
 
             Assert.Equal(5, allowedCount);
             Assert.Equal(8, blockedCount);
-        }
-
-        public ScopeDefinition GetScopeDefinitionOfType(TypeDefinition typeDefinition)
-        {
-            Assert.NotNull(typeDefinition);
-            ScopeDefinition scope = null;
-            NamespaceDefinition currentNamespaceDefinition = typeDefinition.NamespaceDefinition;
-
-            while (scope == null)
-            {
-                Assert.NotNull(currentNamespaceDefinition);
-                scope = currentNamespaceDefinition.ParentScopeOrNamespace as ScopeDefinition;
-                currentNamespaceDefinition = currentNamespaceDefinition.ParentScopeOrNamespace as NamespaceDefinition;
-            }
-
-            return scope;
-        }
-
-        public ScopeReference GetScopeReferenceOfType(TypeReference typeReference)
-        {
-            Assert.NotNull(typeReference);
-            ScopeReference scope = null;
-            NamespaceReference currentNamespaceReference = typeReference.ParentNamespaceOrType as NamespaceReference;
-
-            while (scope == null)
-            {
-                Assert.NotNull(currentNamespaceReference);
-                scope = currentNamespaceReference.ParentScopeOrNamespace as ScopeReference;
-                currentNamespaceReference = currentNamespaceReference.ParentScopeOrNamespace as NamespaceReference;
-            }
-
-            return scope;
-        }
-
-        public void CheckTypeDefinitionForProperWinRTHome(TypeDefinition typeDefinition, string module)
-        {
-            ScopeDefinition scope = GetScopeDefinitionOfType(typeDefinition);
-            Assert.Equal(module, scope.Name.Value);
-            int windowsRuntimeFlag = ((int)System.Reflection.AssemblyContentType.WindowsRuntime << 9);
-            Assert.True((((int)scope.Flags) & windowsRuntimeFlag) == windowsRuntimeFlag);
-        }
-
-
-        public void CheckTypeReferenceForProperWinRTHome(TypeReference typeReference, string module)
-        {
-            ScopeReference scope = GetScopeReferenceOfType(typeReference);
-            Assert.Equal(module, scope.Name.Value);
-            int windowsRuntimeFlag = ((int)System.Reflection.AssemblyContentType.WindowsRuntime << 9);
-            Assert.True((((int)scope.Flags) & windowsRuntimeFlag) == windowsRuntimeFlag);
-        }
-
-        [Fact]
-        public void TestExplicitScopeAttributesForWinRTSingleFilePolicy()
-        {
-            // Test that custom attributes referring to blocked types don't show up in metadata
-
-            var sampleMetadataModule = _context.GetModuleForSimpleName("SampleMetadataAssembly");
-            Cts.MetadataType controlType = _systemModule.GetType("Windows", "Control");
-            Cts.MetadataType derivedFromControl = sampleMetadataModule.GetType("SampleMetadataWinRT", "DerivedFromControl");
-            Cts.MetadataType derivedFromControlInCustomScope = sampleMetadataModule.GetType("SampleMetadataWinRT", "DerivedFromControlAndInCustomScope");
-
-            var policy = new SingleFileMetadataPolicy();
-            policy.Init();
-
-            var transformResult = MetadataTransform.Run(policy,
-                new[] { _systemModule, sampleMetadataModule });
-
-            var controlTypeMetadata = transformResult.GetTransformedTypeDefinition(controlType);
-            var derivedFromControlMetadata = transformResult.GetTransformedTypeDefinition(derivedFromControl);
-            var derivedFromControlInCustomScopeMetadata = transformResult.GetTransformedTypeDefinition(derivedFromControlInCustomScope);
-
-            CheckTypeDefinitionForProperWinRTHome(controlTypeMetadata, "Windows");
-            ScopeDefinition scopeDefOfDerivedFromControlType = GetScopeDefinitionOfType(derivedFromControlMetadata);
-            Assert.Equal("SampleMetadataAssembly", scopeDefOfDerivedFromControlType.Name.Value);
-            CheckTypeDefinitionForProperWinRTHome(derivedFromControlInCustomScopeMetadata, "SampleMetadataWinRT");
-        }
-
-
-        [Fact]
-        public void TestExplicitScopeAttributesForWinRTMultiFilePolicy()
-        {
-            // Test that custom attributes referring to blocked types don't show up in metadata
-
-            var sampleMetadataModule = _context.GetModuleForSimpleName("SampleMetadataAssembly");
-            Cts.MetadataType controlType = _systemModule.GetType("Windows", "Control");
-            Cts.MetadataType derivedFromControl = sampleMetadataModule.GetType("SampleMetadataWinRT", "DerivedFromControl");
-            Cts.MetadataType derivedFromControlInCustomScope = sampleMetadataModule.GetType("SampleMetadataWinRT", "DerivedFromControlAndInCustomScope");
-
-            var policy = new MultifileMetadataPolicy(sampleMetadataModule);
-
-            var transformResult = MetadataTransform.Run(policy,
-                new[] { _systemModule, sampleMetadataModule });
-
-            var controlTypeMetadata = transformResult.GetTransformedTypeReference(controlType);
-            var derivedFromControlMetadata = transformResult.GetTransformedTypeDefinition(derivedFromControl);
-            var derivedFromControlInCustomScopeMetadata = transformResult.GetTransformedTypeDefinition(derivedFromControlInCustomScope);
-
-            CheckTypeReferenceForProperWinRTHome(controlTypeMetadata, "Windows");
-            ScopeDefinition scopeDefOfDerivedFromControlType = GetScopeDefinitionOfType(derivedFromControlMetadata);
-            Assert.Equal("SampleMetadataAssembly", scopeDefOfDerivedFromControlType.Name.Value);
-            CheckTypeDefinitionForProperWinRTHome(derivedFromControlInCustomScopeMetadata, "SampleMetadataWinRT");
         }
     }
 }
