+++ src/ILCompiler.MetadataTransform/tests/WindowsWinrtMetadataAssembly/WindowsWinrtMetadata.cs
@@ -1,0 +1,25 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Runtime.CompilerServices;
+
+#pragma warning disable 414
+#pragma warning disable 67
+#pragma warning disable 3009
+#pragma warning disable 3016
+#pragma warning disable 3001
+#pragma warning disable 3015
+#pragma warning disable 169
+#pragma warning disable 649
+
+namespace Windows
+{
+    [Internal.Reflection.ExplicitScope("Windows, Version=255.255.255.255, Culture=neutral, PublicKeyToken=null, ContentType=WindowsRuntime")]
+    public class Control
+    {
+    }
+}
